+++ SBBMobileDesignSystemSwiftUI/Colors/SBBColorName.swift
@@ -1,0 +1,53 @@
+//
+//  Copyright © 2020 SBB. All rights reserved.
+//
+
+import Foundation
+
+public enum SBBColorName: String {
+    // Colors
+    case red = "Red"
+    case red125 = "Red125"
+    case red150 = "Red150"
+    case white = "White"
+    case milk = "Milk"
+    case cloud = "Cloud"
+    case silver = "Silver"
+    case aluminium = "Aluminium"
+    case platinum = "Platinum"
+    case cement = "Cement"
+    case graphite = "Graphite"
+    case storm = "Storm"
+    case smoke = "Smoke"
+    case metal = "Metal"
+    case granite = "Granite"
+    case anthracite = "Anthracite"
+    case iron = "Iron"
+    case black = "Black"
+    case blue = "Blue"
+    case sky = "Sky"
+    case night = "Night"
+    case violet = "Violet"
+    case autumn = "Autumn"
+    case orange = "Orange"
+    case peach = "Peach"
+    case lemon = "Lemon"
+    case green = "Green"
+    case shadow = "Shadow"
+    case charcoal = "Charcoal"
+    
+    // Semantic Colors
+    case textBlack = "textBlack"
+    case textMetal = "textMetal"
+    case textRed = "textRed"
+    case textWhite = "textWhite"
+    case background = "background"
+    case buttonBackground = "buttonBackground"
+    case disabledButtonBackground = "disabledButtonBackground"
+    case viewBackground = "viewBackground"
+    case divider = "divider"
+    case border = "border"
+    case tabViewBackground = "tabViewBackground"
+    case placeholder = "placeholder"
+    
+}