+++ SBBMobileDesignSystemSwiftUI/Views/SBBPaginationView.swift
@@ -20,11 +20,11 @@ public struct SBBPaginationView: View {
                 HStack {    // Workaround needed, so that the Circle-Views get reloaded after a currentPageIndex change, see https://developer.apple.com/forums/thread/131577
                     if index == self.currentPageIndex {
                         Circle()
-                            .fill(SBBColor.textBlack)
+                            .fill(Color.sbbColor(.textBlack))
                             .frame(width: 8, height: 8)
                     } else {
                         Circle()
-                            .fill(SBBColor.paginationInactive)
+                            .fill(Color.sbbColorInternal(.paginationInactive))
                             .frame(width: 4, height: 4)
                     }
                 }