+++ SBBMobileDesignSystemSwiftUIDemo/UI/DetailViews/InfoViewCollapsibleDemo.swift
@@ -39,7 +39,7 @@ struct InfoViewCollapsibleDemo: View {
             .padding(16)
         }
             .navigationBarTitle("InfoViewCollapsible")
-            .background(SBBColor.background.edgesIgnoringSafeArea(.bottom))
+            .background(Color.sbbColor(.background).edgesIgnoringSafeArea(.bottom))
             .colorScheme(colorScheme)
     }
 }