+++ SBBMobileDesignSystemSwiftUIDemo/UI/DetailViews/ListItemDemo.swift
@@ -43,7 +43,7 @@ struct ListItemDemo: View {
             .padding(16)
         }
             .navigationBarTitle("ListItem")
-            .background(SBBColor.background.edgesIgnoringSafeArea(.bottom))
+            .background(Color.sbbColor(.background).edgesIgnoringSafeArea(.bottom))
             .colorScheme(colorScheme)
     }
 }